+++ docs/examples/datepicker/main.tsx
@@ -6,15 +6,15 @@ export default function Example() {
   const [dateValue, setDateValue] = useState<Date | null>(new Date(1985, 6, 4));
 
   return (
-    <Flex alignItems="start" height="100%" justifyContent="center" width="100%">
-      <Box padding={2}>
+    <Box padding={8} width="100%">
+      <Flex alignItems="start" height="100%" justifyContent="center" width="100%">
         <DatePicker
           id="main"
           label="Select a date"
           onChange={({ value }) => setDateValue(value)}
           value={dateValue}
         />
-      </Box>
-    </Flex>
+      </Flex>{' '}
+    </Box>
   );
 }